SmartResume-Generator/requirements.txt
streamlit>=1.20.0
google-generativeai>=0.1.0
python-docx>=0.8.11
fpdf>=1.7.2
python-dotenv>=1.0.0


# to run use this cmd:   python -m streamlit run app.py 